src/audiobooks/requirements.txt
requests
mcp
google-api-python-client
google-auth-oauthlib
pydub
imageio-ffmpeg
coqui-tts
torch
torchaudio
google-genai
pymupdf
beautifulsoup4
sentence-transformers
pytest
pytest-mock
requests-mock
